Fix incorrect type of id in CrudController and ajust tests (#20)

+++ src/main/java/controller/CRUD.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 public interface CRUD<T> {
     void create(T record) throws SQLException;
-    Optional<T> read(long id) throws SQLException;
-    void update(long id, T record) throws SQLException;
-    void delete(long id) throws SQLException;
+    Optional<T> read(int id) throws SQLException;
+    void update(int id, T record) throws SQLException;
+    void delete(int id) throws SQLException;
 }
